Merge From S09P31B109-279/Test/BE/API테스트및오류수정 into BackEnd

See merge request s09-final/S09P31B109!58

+++ BackEnd/src/main/java/com/ssafy/bangrang/global/fcm/service/AlarmServieImpl.java
@@ -90,7 +90,7 @@ public class AlarmServieImpl implements AlarmService {
                     .content(alarm.getContent())
                     .eventIdx(eventIdx)
                     .alarmCreatedDate(alarm.getCreatedDate())
-                    .alarmStatus(alarm.getStatus())
+                    .alarmStatus(alarm.getStatus().intValue())
                     .build();
 
             result.add(alarmListResponseDto);
